ldn: Adjust namespaces
